fix: 修复ffmpeg文件引用问题导致去重失败
